zzfatorar: Incluindo o comando factor, quando presente.

+++ zz/zzfatorar.sh
@@ -1,5 +1,5 @@
 # ----------------------------------------------------------------------------
-# http://www.primos.mat.br/primeiros_10000_primos.txt
+# http://www.primos.mat.br
 # Fatora um número em fatores primos.
 # Com as opções:
 #   --atualiza: atualiza o cache com 10 mil primos ( padrão e rápida ).
@@ -16,7 +16,7 @@
 #
 # Autor: Itamar <itamarnet (a) yahoo com br>
 # Desde: 2013-03-14
-# Versão: 2
+# Versão: 3
 # Licença: GPL
 # Requisitos: zzjuntalinhas zzdos2unix
 # ----------------------------------------------------------------------------
@@ -45,15 +45,15 @@ zzfatorar ()
 			# Atualiza o cache com uma listagem com 1 milhão de números primos.
 			# É um processo bem mais lento, devendo ser usado quando o cache normal não atende.
 			rm -f "$cache"
-			if type 7z >/dev/null 2>&1
+			if type 7z >/dev/null 2>&1 && ! type factor >/dev/null 2>&1
 			then
 				zztool eco "Atualizando cache."
 				wget -q http://www.primos.mat.br/dados/50M_part1.7z -O /tmp/primos.7z
-				7z e /tmp/primos.7z 2>&1 >/dev/null
+				7z e /tmp/primos.7z >/dev/null 2>&1
 				rm -f /tmp/primos.7z
 				awk '{for(i=1;i<=NF;i++) print $i }' 50M_part1.txt > "$cache"
 				rm -f 50M_part1.txt
-				zzdos2unix "$cache" 2>&1 >/dev/null
+				zzdos2unix "$cache" >/dev/null 2>&1
 				zztool eco "Cache atualizado."
 			fi
 			shift
@@ -72,15 +72,21 @@ zzfatorar ()
 		esac
 	done
 
-	# Se o cache está vazio, baixa listagem da Internet
-	if ! test -s "$cache"
-	then
-		$ZZWWWDUMP "$url" | awk '{for(i=1;i<=NF;i++) print $i }' > "$cache"
-	fi
-
 	# Apenas para numeros inteiros
 	if zztool testa_numero "$1" && test $1 -ge 2
 	then
+
+		if type factor >/dev/null 2>&1
+		then
+			# Se existe o camando factor usa-o
+			factor $1 | sed 's/.*: //g' | awk '{for(i=1;i<=NF;i++) print $i }' | uniq > "$cache"
+			primo_atual=$(head -n 1 "$cache")
+		elif ! test -s "$cache"
+		then
+			# Se o cache está vazio, baixa listagem da Internet
+			$ZZWWWDUMP "$url" | awk '{for(i=1;i<=NF;i++) print $i }' > "$cache"
+		fi
+
 		# Se o número fornecido for primo, retorna-o e sai
 		grep "^${1}$" ${cache} > /dev/null
 		[ "$?" = "0" ] && { echo " $1 é um número primo."; return; }
@@ -88,7 +94,7 @@ zzfatorar ()
 		num_atual="$1"
 		tamanho=$((${#1} + 1))
 
-		# Enquanto a resultado for maior que o número primo continua, ou dentro das 10000 primos listados
+		# Enquanto a resultado for maior que o número primo continua, ou dentro dos primos listados no cache.
 		while [ ${num_atual} -gt ${primo_atual} -a ${linha_atual} -le $(wc -l "$cache" | tr -d -c '[0-9]') ]
 		do
 
@@ -115,9 +121,12 @@ zzfatorar ()
 			fi
 
 			# Definindo o número primo a ser usado
-			linha_atual=$((${linha_atual} + 1))
-			primo_atual=$(sed -n "${linha_atual}p" "$cache")
-			[ ${#primo_atual} -eq 0 ] && { zztool eco " Valor não fatorável nessa configuração do script!"; return 1; }
+			if [ "${num_atual}" != "1" ]
+			then
+				linha_atual=$((${linha_atual} + 1))
+				primo_atual=$(sed -n "${linha_atual}p" "$cache")
+				[ ${#primo_atual} -eq 0 ] && { zztool eco " Valor não fatorável nessa configuração do script!"; return 1; }
+			fi
 		done
 
 		if [ "$bc" != "2" ]
